backend/requirements.dev.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
redis
pytest
httpx
email-validator
PyJWT
pydantic-settings
bcrypt
